RSGISLib function popImageStats appears to build pyramids as well, so removing steps, also moving some files / renamed to make them more sensible
